backend/requirements.txt
flask==3.0.2
Flask-Cors==4.0.0
numpy==1.26.4
opencv-python==4.9.0.80
pillow==10.2.0